Fixed CPU/RAM hogging with large response text.
The original layout had the editor control auto extend to show the entire
response text, then use a scroll viewer for the entire right panel. It is
a nice UI (similar to that of Postman), but it means the editor control
have to render every line of text. This becomes very CPU and RAM intensive
when dealing with multi-megabyte responses (e.g. herd animals in
itops.mobile, bulls in itops.datamate).

This PR changes the layout of the controls to limit the editor to the size
of the window, in order to leverage the virtualization implemented by
AvaloneEdit control (i.e. only the text visible is syntax-highlighted and
rendered).