History delete button layout change
As per discussed in #37, horizontal scrolling is now disabled in the
history listbox, which removed the requirement for dynamic delete button
positioning. This cleared up a lot of complex hoops that was implemented
for the sole reason of making the delete button being snapped to the edge
of the grid splitter rather than hidden underneath when scroll width is
too wide.

This also removed requirement for custom MaloneListBox.